Added check box and completed Create User code
There is now a check box that resides under the second password box (the confirmation password box) to display the passwords to the user (unset and set the echo character for the password fields). The Create User code has been completed and will store a secure version of the users password, their username and the respective salt for their password in their user information data object.