Update Record.java
Modified the original consturctor for a record object as a majority of the data is not captured when initially creating the account. Created an additional constructor to the a char array as the password object and the constructor creates a string builder variable, concatenates all the characters in the char array and sets the password variable to the value of the string builder as a string.