Update NewAccount.java
Added a variable to store a reference to the editting record. Created an additional "setReferences" method so that the current editting record could also be passed in as well as the parent menu and logged in user. Correctly checking for duplicate account names even if editting a record -> the current editting records name will be omitted from the check for duplicate names. Modified the create method to include the posibility of editting a record and saving this data instead of creating a new one.