See description
Corrected variable naming convention error. Added a method to toggle the lock state of the record and reflect this by locking the coordinate boxes. Added a method to reset the menu components separately. As list items are selected, check for the lock state and reflect this in the main menu. Mthod to check, adjust components and set the lock state of a record.